fix service specific to email_service & restore navigation & enforce policy
