+++ plugins/email_service/app/controllers/email_service/application_controller.rb
@@ -5,9 +5,12 @@ module EmailService
     include AwsSesHelper
     include EmailHelper
     include TemplateHelper
-    include VerificationHelper
+    include VerificationsHelper
     include ConfigsetHelper
 
+    authorization_context 'email_service'
+    authorization_required
+    
     def ui_switcher
       if current_user.has_role?('cloud_support_tools_viewer')
         redirect_to emails_path
