+++ plugins/email_service/app/controllers/email_service/settings_controller.rb
@@ -2,6 +2,9 @@ module EmailService
   class SettingsController < ::EmailService::ApplicationController
     before_action :restrict_access
 
+    authorization_context 'email_service'
+    authorization_required
+    
     def index
       @cronus_activated = false
       creds = get_ec2_creds
