+++ plugins/email_service/app/controllers/email_service/web_controller.rb
@@ -2,6 +2,9 @@ module EmailService
   class WebController < ::EmailService::ApplicationController
     before_action :ui_switcher
 
+    authorization_context 'email_service'
+    authorization_required
+
     def index
     end
