+++ plugins/email_service/app/helpers/email_service/templated_email_helper.rb
@@ -1,5 +1,8 @@
 module EmailService
   module TemplatedEmailHelper
+
+    include TemplateHelper
+
     class TemplatedEmail
       VALID_EMAIL_REGEX = /\A[\w+\-.]+@[a-z\d\-.]+\.[a-z]+\z/i
