+++ plugins/email_service/app/services/service_layer/email_service_service.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+module ServiceLayer
+  # This class implements the email_service api
+  class EmailServiceService < Core::ServiceLayer::Service
+    include EmailServiceServices::Email
+    include EmailServiceServices::Template
+    # include EmailServiceServices::VerifiedEmail
+
+    def available?(_action_name_sym = nil)
+      elektron.service?('email-aws')
+    end
+
+    def elektron_email_service
+      @elektron_email_service ||= elektron.service(
+        'identity', path_prefix: '/v3'
+      )
+    end
+
+    # def elektron_identity_service
+    #   @elektron_identity_service ||= elektron.service(
+    #     'identity', path_prefix: '/v3'
+    #   )
+    # end
+
+  end
+end
