+++ plugins/email_service/config/policy.json
@@ -2,32 +2,30 @@
   "email_service_user": "role:email_user",
   "email_service_admin": "role:email_admin",
   "email_role_required": "role:email_user or role:email_admin",
-  "email_service:email_create": "rule:email_role_required",
+  "email_service:emails_create": "rule:email_role_required",
+  "email_service:emails_stats": "rule:email_role_required",
+  "email_service:emails_info": "rule:email_role_required",
   "email_service:templated_email_create": "rule:email_role_required",
-  "email_service:configset_list": "rule:email_role_required",
-  "email_service:configset_get": "rule:email_role_required",
-  "email_service:configset_create": "rule:email_role_required",
-  "email_service:configset_delete": "rule:email_role_required",
-  "email_service:configset_update": "rule:email_role_required",
-  "email_service:settings_list": "rule:email_role_required",
-  "email_service:settings_get": "rule:email_role_required",
-  "email_service:settings_create": "rule:email_service_admin",
-  "email_service:settings_delete": "rule:email_service_admin",
-  "email_service:statistics_list": "rule:email_role_required",
-  "email_service:verified_email_list": "rule:email_role_required",
-  "email_service:verified_email_get": "rule:email_role_required",
-  "email_service:verified_email_create": "rule:email_role_required",
-  "email_service:verified_email_delete": "rule:email_role_required",
-  "email_service:verified_domain_list": "rule:email_role_required",
-  "email_service:verified_domain_get": "rule:email_role_required",
-  "email_service:verified_domain_create": "rule:email_role_required",
-  "email_service:verified_domain_delete": "rule:email_role_required",
+  "email_service:configset_new_configset": "rule:email_role_required",
+  "email_service:configset_show_configset": "rule:email_role_required",
+  "email_service:configset_create_configset": "rule:email_role_required",
+  "email_service:configset_destroy_configset": "rule:email_role_required",
+  "email_service:settings_show_config": "rule:email_role_required",
+  "email_service:settings_enable_cronus": "rule:email_service_admin",
+  "email_service:settings_disable_cronus": "rule:email_service_admin",
   "email_service:template_list": "rule:email_role_required",
   "email_service:template_get": "rule:email_role_required",
   "email_service:template_create": "rule:email_role_required",
   "email_service:template_delete": "rule:email_role_required",
   "email_service:template_update": "rule:email_role_required",
-  "email_service:email_stats": "rule:email_role_required",
-  "email_service:email_info": "rule:email_role_required",
-  "email_service:web_list": "rule:email_role_required"
+  "email_service:settings": "rule:email_role_required",
+  "email_service:verifications_create": "rule:email_role_required",
+  "email_service:verifications_delete": "rule:email_role_required",
+  "email_service:domain_verifications_create": "rule:email_role_required",
+  "email_service:domain_verifications_delete": "rule:email_role_required",
+  "email_service:domain_verifications_verify_dkim": "rule:email_role_required",
+  "email_service:domain_verifications_activate_dkim": "rule:email_role_required",
+  "email_service:domain_verifications_deactivate_dkim": "rule:email_role_required",
+  "email_service:settings_create": "rule:email_service_admin",
+  "email_service:settings_delete": "rule:email_service_admin"
 }
