--- plugins/identity/app/services/service_layer/identity_services/aws/template.rb
@@ -1,51 +1,0 @@
-# frozen_String_literal: true
-
-module ServiceLayer
-  module IdentityServices
-    module AWS
-      # template api implementation
-      module Template
-        def template_map
-          @template_map ||= class_map_proc(::EmailService::Template)
-        end
-
-        def new_template(attributes = {})
-          template_map.call(attributes)
-        end
-
-        def templates(filter = {})
-          response = elektron_identity.get('templates', filter)
-          {
-            items: response.map_to('body.templates', &template_map),
-            total: response.body['total']
-          }
-        end
-
-        def find_template!(uuid)
-          elektron_identity.get("templates/#{uuid}").map_to(
-            'body', &template_map
-          )
-        end
-
-        def find_template(uuid)
-          find_template!(uuid)
-        rescue Elektron::Errors::ApiResponse => _e
-          nil
-        end
-
-        ################ MODEL/API INTERFACE ###################
-        def create_template(attributes = {})
-          template_ref = elektron_identity.post('templates') do
-            attributes
-          end.body
-          attributes.merge(template_ref)
-        end
-
-        def delete_template(id)
-          elektron_identity.delete("templates/#{id}")
-        end
-        
-      end
-    end
-  end
-end
